Switch to 32-bit types as specified by SEGGER
Closes #1013

+++ firmware/defmt-rtt/src/channel.rs
@@ -1,6 +1,6 @@
 use core::{
     ptr,
-    sync::atomic::{AtomicUsize, Ordering},
+    sync::atomic::{AtomicU32, Ordering},
 };
 
 use crate::{consts::BUF_SIZE, MODE_BLOCK_IF_FULL, MODE_MASK};
@@ -8,18 +8,20 @@ use crate::{consts::BUF_SIZE, MODE_BLOCK_IF_FULL, MODE_MASK};
 /// RTT Up channel
 #[repr(C)]
 pub(crate) struct Channel {
+    /// Name of the channel (null terminated)
     pub name: *const u8,
     /// Pointer to the RTT buffer.
     pub buffer: *mut u8,
-    pub size: usize,
+    /// Size, in bytes, of the RTT buffer
+    pub size: u32,
     /// Written by the target.
-    pub write: AtomicUsize,
+    pub write: AtomicU32,
     /// Written by the host.
-    pub read: AtomicUsize,
+    pub read: AtomicU32,
     /// Channel properties.
     ///
     /// Currently, only the lowest 2 bits are used to set the channel mode (see constants below).
-    pub flags: AtomicUsize,
+    pub flags: AtomicU32,
 }
 
 impl Channel {
@@ -46,8 +48,8 @@ impl Channel {
         }
 
         // calculate how much space is left in the buffer
-        let read = self.read.load(Ordering::Relaxed);
-        let write = self.write.load(Ordering::Acquire);
+        let read = self.read.load(Ordering::Relaxed) as usize;
+        let write = self.write.load(Ordering::Acquire) as usize;
         let available = available_buffer_size(read, write);
 
         // abort if buffer is full
@@ -59,7 +61,7 @@ impl Channel {
     }
 
     fn nonblocking_write(&self, bytes: &[u8]) -> usize {
-        let write = self.write.load(Ordering::Acquire);
+        let write = self.write.load(Ordering::Acquire) as usize;
 
         // NOTE truncate at BUF_SIZE to avoid more than one "wrap-around" in a single `write` call
         self.write_impl(bytes, write, BUF_SIZE)
@@ -82,8 +84,10 @@ impl Channel {
         }
 
         // adjust the write pointer, so the host knows that there is new data
-        self.write
-            .store(cursor.wrapping_add(len) % BUF_SIZE, Ordering::Release);
+        self.write.store(
+            (cursor.wrapping_add(len) % BUF_SIZE) as u32,
+            Ordering::Release,
+        );
 
         // return the number of bytes written
         len
